requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Werkzeug==3.1.3
PyJWT==2.8.0
qrcode[pil]==7.4.2
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
